runs on -Wall ...

+++ 2-selection_sort.c
@@ -22,7 +22,7 @@ void swap(int *array, int a, int b)
  */
 void selection_sort(int *array, size_t size)
 {
-	int i, j, min;
+	size_t i, j, min;
 
 	for (i = 0; i < size - 1; i++)
 	{
